270824: Move OurValues into its own component
